+++ frontend/src/components/@shared/Dropdown/index.tsx
@@ -2,7 +2,7 @@ import { RefObject, useEffect } from "react";
 
 import useDropdown from "@src/components/@shared/Dropdown/@hooks/useDropdown";
 
-import useOuterClick from "@src/hooks/useOuterClick";
+import useOuterClick from "@src/hooks/@shared/useOuterClick";
 
 interface ChildrenProps {
   innerRef: RefObject<HTMLDivElement>;
