+++ frontend/src/components/AddReminder/@hooks/useDatePicker.ts
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState } from "react";
 
 import { parseDate } from "@src/components/AddReminder/@utils";
 
-import useInput from "@src/hooks/useInput";
+import useInput from "@src/hooks/@shared/useInput";
 
 import { PICKER_OPTION_SCROLL } from "@src/@constants";
 import { getDateInformation } from "@src/@utils";
