+++ frontend/src/components/AddReminder/@hooks/useTimePicker.ts
@@ -6,7 +6,7 @@ import {
   parseTime,
 } from "@src/components/AddReminder/@utils";
 
-import useInput from "@src/hooks/useInput";
+import useInput from "@src/hooks/@shared/useInput";
 
 import { PICKER_OPTION_SCROLL } from "@src/@constants";
 import { getDateInformation, getMeridiemTime } from "@src/@utils";
