+++ frontend/src/components/SearchForm/@hooks/useSelectChannels.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useState } from "react";
 
-import useGetSubscribedChannels from "@src/hooks/query/useGetSubscribedChannels";
+import useGetSubscribedChannels from "@src/hooks/@query/useGetSubscribedChannels";
 
 import { SubscribedChannel } from "@src/@types/shared";
