+++ frontend/src/hooks/@shared/useOuterClick.ts
