+++ frontend/src/hooks/useRecentFeedPath.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 
-import useWebStorage, { STORAGE_KIND } from "@src/hooks/useWebStorage";
+import useWebStorage, { STORAGE_KIND } from "@src/hooks/@shared/useWebStorage";
 
 function useRecentFeedPath() {
   const { key, pathname } = useLocation();
