+++ frontend/src/pages/Bookmark/index.tsx
@@ -6,9 +6,9 @@ import MessageCard from "@src/components/MessageCard";
 import BookmarkButton from "@src/components/MessageCard/MessageIconButtons/BookmarkButton";
 import MessagesLoadingStatus from "@src/components/MessageCard/MessagesLoadingStatus";
 
-import useGetInfiniteBookmarks from "@src/hooks/query/useGetInfiniteBookmarks";
-import useMutateBookmark from "@src/hooks/query/useMutateBookmark";
-import useScrollToTop from "@src/hooks/useScrollToTop";
+import useGetInfiniteBookmarks from "@src/hooks/@query/useGetInfiniteBookmarks";
+import useMutateBookmark from "@src/hooks/@query/useMutateBookmark";
+import useScrollToTop from "@src/hooks/@shared/useScrollToTop";
 
 import { FlexColumn } from "@src/@styles/shared";
 import { extractResponseBookmarks, parseMeridemTime } from "@src/@utils";
